1/requirements.txt
pyyaml>=6.0
typing-extensions>=4.5.0
requests>=2.28.0
flask>=2.0.0
typing-extensions>=4.5.0
mysql-connector-python>=8.0.0
pymysql>=1.0.0
beautifulsoup4
jsonpath-ng
markdown
sqlalchemy
pymysql
psycopg2-binary
lxml
pyyaml
jinja2
pandas
# pip install -r requirements.txt
# pip install requests beautifulsoup4 markdown jsonpath-ng
# pip install sqlalchemy pymysql psycopg2-binary lxml pyyaml jinja2